move custom argparse action to util
